Create TimelineCardWidget
