+++ pubspec.yaml
@@ -28,7 +28,6 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-  provider: ^5.0.0
   video_player: ^2.1.6
 
   firebase_analytics: ^8.1.2
@@ -36,6 +35,9 @@ dependencies:
   cloud_firestore: ^2.3.0
   firebase_storage: ^9.0.0
   firebase_core: ^1.3.0
+
+  flutter_hooks: ^0.18.0
+  hooks_riverpod: ^1.0.0-dev.6
 
 dev_dependencies:
   flutter_test:
